refactor: Use a single file for all utils
