+++ src/components/CourseCard/CourseCard.tsx
@@ -1,8 +1,8 @@
 import { Course } from "@/types/entities/general-entities";
-import { getCourseInitials } from "@/utils/get-course-initials";
 import { Link } from "react-router-dom";
 
 import { CourseDropDown } from "./CourseDropDown";
+import { getCourseInitials } from "@/utils/utils";
 
 interface CourseCardProps {
   course: Course;
