+++ src/components/CourseCard/CourseDropDown.tsx
@@ -13,7 +13,7 @@ import { SessionRole } from "@/hooks/useSession";
 import { getInvitationCodeService } from "@/services/courses/get-invitation-code.service";
 import { toggleCourseVisibilityService } from "@/services/courses/toggle-course-visibility.service";
 import { Course } from "@/types/entities/general-entities";
-import { copyToClipboard } from "@/utils/copy-to-clipboard";
+import { copyToClipboard } from "@/utils/utils";
 import {
   ClipboardCopy,
   Eye,
