--- src/utils/copy-to-clipboard.ts
@@ -1,8 +1,0 @@
-export const copyToClipboard = async (text: string): Promise<boolean> => {
-  try {
-    navigator.clipboard.writeText(text);
-    return true;
-  } catch {
-    return false;
-  }
-};
